Added server folder, cleaned up some code in app.js files. Moved server side code into server folder.
